Fix printf warnings

+++ example_programs/websocket_example/websocket_client.c
@@ -97,7 +97,7 @@ void websocket_incoming_message_callback (const struct _u_request * request,
   }
   y_log_message(Y_LOG_LEVEL_DEBUG, "Incoming message, opcode: 0x%02x, mask: %d, len: %zu", last_message->opcode, last_message->has_mask, last_message->data_len);
   if (last_message->opcode == U_WEBSOCKET_OPCODE_TEXT) {
-    y_log_message(Y_LOG_LEVEL_DEBUG, "text payload '%.*s'", last_message->data_len, last_message->data);
+    y_log_message(Y_LOG_LEVEL_DEBUG, "text payload '%.*s'", (int)last_message->data_len, last_message->data);
   } else if (last_message->opcode == U_WEBSOCKET_OPCODE_BINARY) {
     y_log_message(Y_LOG_LEVEL_DEBUG, "binary payload");
   }
